\name{A061105}
\alias{A061105}
\alias{A061105}
\title{Example Wheel Running Data...}
\description{Example Wheel Running Data
Wheel Running Data for 2006-11-05 for Computers A, B, C, and D}
\alias{A061105}
\alias{B061105}
\alias{C061105}
\alias{D061105}
\docType{data}
\keyword{datasets}
\examples{data(A061105)
str(A061105)}

