\name{lineaggr}
\alias{lineaggr}
\title{Aggregate by Line}
\usage{lineaggr(dat)
}
\description{Aggregate by Line}
\details{Aggregate running data by line}
\value{Data aggregated by line}
\author{Kevin Middleton (kmm@csusb.edu)}
\seealso{\code{\link{linetypeaggr}}}
\keyword{univar}
\arguments{\item{dat}{A data.frame}
}

