\name{plot_running}
\alias{plot_running}
\title{Plot wheel running (new system)}
\usage{
  plot_running(data, save = FALSE, start.at.0 = TRUE)
}
\arguments{
  \item{data}{Data.frame returned from
  \code{\link{analyze_running}}.}

  \item{save}{Boolean (default = FALSE) Should plots be
  saved?}

  \item{start.at.0}{Boolean (default = TRUE) Should y-axis
  start at 0?}
}
\description{
  Plot wheel running (new system)
}
\details{
  Create diagnostic plots for wheel running data,
}
\author{
  Kevin Middleton
}

