\name{stackdata}
\alias{stackdata}
\title{Stack Data Frames}
\usage{stackdata(...)
}
\description{Stack data frames on top of one another.}
\details{Take an arbitrary number of aggregated data.frames in and return them as a
single data.frame.}
\value{A data.frame with \dots stacked.}
\author{Kevin Middleton (kmm@csusb.edu)}
\seealso{\code{\link{rbind}}}
\keyword{data}
\arguments{\item{\dots}{data.frames to \code{rbind}}
}

