\docType{data}
\name{A061105}
\alias{A061105}
\alias{B061105}
\alias{C061105}
\alias{D061105}
\title{Example Wheel Running Data
Wheel Running Data for 2006-11-05 for Computers A, B, C, and D}
\format{A data frame with 1385 observations on the following 157 variables.
\describe{
\item{list("V1")}{a numeric vector}
\item{list("V2")}{a numeric vector}
\item{list("V3")}{a numeric vector}
\item{list("V4")}{a numeric vector}
\item{list("V5")}{a numeric vector}
\item{list("V6")}{a numeric vector}
\item{list("V7")}{a numeric vector}
\item{list("V8")}{a numeric vector}
\item{list("V9")}{a numeric vector}
\item{list("V10")}{a numeric vector}
\item{list("V11")}{a numeric vector}
\item{list("V12")}{a numeric vector}
\item{list("V13")}{a numeric vector}
\item{list("V14")}{a numeric vector}
\item{list("V15")}{a numeric vector}
\item{list("V16")}{a numeric vector}
\item{list("V17")}{a numeric vector}
\item{list("V18")}{a numeric vector}
\item{list("V19")}{a numeric vector}
\item{list("V20")}{a numeric vector}
\item{list("V21")}{a numeric vector}
\item{list("V22")}{a numeric vector}
\item{list("V23")}{a numeric vector}
\item{list("V24")}{a numeric vector}
\item{list("V25")}{a numeric vector}
\item{list("V26")}{a numeric vector}
\item{list("V27")}{a numeric vector}
\item{list("V28")}{a numeric vector}
\item{list("V29")}{a numeric vector}
\item{list("V30")}{a numeric vector}
\item{list("V31")}{a numeric vector}
\item{list("V32")}{a numeric vector}
\item{list("V33")}{a numeric vector}
\item{list("V34")}{a numeric vector}
\item{list("V35")}{a numeric vector}
\item{list("V36")}{a numeric vector}
\item{list("V37")}{a numeric vector}
\item{list("V38")}{a numeric vector}
\item{list("V39")}{a numeric vector}
\item{list("V40")}{a numeric vector}
\item{list("V41")}{a numeric vector}
\item{list("V42")}{a numeric vector}
\item{list("V43")}{a numeric vector}
\item{list("V44")}{a numeric vector}
\item{list("V45")}{a numeric vector}
\item{list("V46")}{a numeric vector}
\item{list("V47")}{a numeric vector}
\item{list("V48")}{a numeric vector}
\item{list("V49")}{a numeric vector}
\item{list("V50")}{a numeric vector}
\item{list("V51")}{a numeric vector}
\item{list("V52")}{a numeric vector}
\item{list("V53")}{a numeric vector}
\item{list("V54")}{a numeric vector}
\item{list("V55")}{a numeric vector}
\item{list("V56")}{a numeric vector}
\item{list("V57")}{a numeric vector}
\item{list("V58")}{a numeric vector}
\item{list("V59")}{a numeric vector}
\item{list("V60")}{a numeric vector}
\item{list("V61")}{a numeric vector}
\item{list("V62")}{a numeric vector}
\item{list("V63")}{a numeric vector}
\item{list("V64")}{a numeric vector}
\item{list("V65")}{a numeric vector}
\item{list("V66")}{a numeric vector}
\item{list("V67")}{a numeric vector}
\item{list("V68")}{a numeric vector}
\item{list("V69")}{a numeric vector}
\item{list("V70")}{a numeric vector}
\item{list("V71")}{a numeric vector}
\item{list("V72")}{a numeric vector}
\item{list("V73")}{a numeric vector}
\item{list("V74")}{a numeric vector}
\item{list("V75")}{a numeric vector}
\item{list("V76")}{a numeric vector}
\item{list("V77")}{a numeric vector}
\item{list("V78")}{a numeric vector}
\item{list("V79")}{a numeric vector}
\item{list("V80")}{a numeric vector}
\item{list("V81")}{a numeric vector}
\item{list("V82")}{a numeric vector}
\item{list("V83")}{a numeric vector}
\item{list("V84")}{a numeric vector}
\item{list("V85")}{a numeric vector}
\item{list("V86")}{a numeric vector}
\item{list("V87")}{a numeric vector}
\item{list("V88")}{a numeric vector}
\item{list("V89")}{a numeric vector}
\item{list("V90")}{a numeric vector}
\item{list("V91")}{a numeric vector}
\item{list("V92")}{a numeric vector}
\item{list("V93")}{a numeric vector}
\item{list("V94")}{a numeric vector}
\item{list("V95")}{a numeric vector}
\item{list("V96")}{a numeric vector}
\item{list("V97")}{a numeric vector}
\item{list("V98")}{a numeric vector}
\item{list("V99")}{a numeric vector}
\item{list("V100")}{a numeric vector}
\item{list("V101")}{a numeric vector}
\item{list("V102")}{a numeric vector}
\item{list("V103")}{a numeric vector}
\item{list("V104")}{a numeric vector}
\item{list("V105")}{a numeric vector}
\item{list("V106")}{a numeric vector}
\item{list("V107")}{a numeric vector}
\item{list("V108")}{a numeric vector}
\item{list("V109")}{a numeric vector}
\item{list("V110")}{a numeric vector}
\item{list("V111")}{a numeric vector}
\item{list("V112")}{a numeric vector}
\item{list("V113")}{a numeric vector}
\item{list("V114")}{a numeric vector}
\item{list("V115")}{a numeric vector}
\item{list("V116")}{a numeric vector}
\item{list("V117")}{a numeric vector}
\item{list("V118")}{a numeric vector}
\item{list("V119")}{a numeric vector}
\item{list("V120")}{a numeric vector}
\item{list("V121")}{a numeric vector}
\item{list("V122")}{a numeric vector}
\item{list("V123")}{a numeric vector}
\item{list("V124")}{a numeric vector}
\item{list("V125")}{a numeric vector}
\item{list("V126")}{a numeric vector}
\item{list("V127")}{a numeric vector}
\item{list("V128")}{a numeric vector}
\item{list("V129")}{a numeric vector}
\item{list("V130")}{a numeric vector}
\item{list("V131")}{a numeric vector}
\item{list("V132")}{a numeric vector}
\item{list("V133")}{a numeric vector}
\item{list("V134")}{a numeric vector}
\item{list("V135")}{a numeric vector}
\item{list("V136")}{a numeric vector}
\item{list("V137")}{a numeric vector}
\item{list("V138")}{a numeric vector}
\item{list("V139")}{a numeric vector}
\item{list("V140")}{a numeric vector}
\item{list("V141")}{a numeric vector}
\item{list("V142")}{a numeric vector}
\item{list("V143")}{a numeric vector}
\item{list("V144")}{a numeric vector}
\item{list("V145")}{a numeric vector}
\item{list("V146")}{a numeric vector}
\item{list("V147")}{a numeric vector}
\item{list("V148")}{a numeric vector}
\item{list("V149")}{a numeric vector}
\item{list("V150")}{a numeric vector}
\item{list("V151")}{a numeric vector}
\item{list("V152")}{a numeric vector}
\item{list("V153")}{a numeric vector}
\item{list("V154")}{a numeric vector}
\item{list("V155")}{a numeric vector}
\item{list("V156")}{a numeric vector}
\item{list("V157")}{a numeric vector}
}}
\description{
  Example Wheel Running Data Wheel Running Data for
  2006-11-05 for Computers A, B, C, and D
}
\examples{
data(A061105)
str(A061105)
}
\keyword{datasets}

