\docType{data}
\name{A061105}
\alias{A061105}
\alias{B061105}
\alias{C061105}
\alias{D061105}
\title{Example Wheel Running Data
Wheel Running Data for 2006-11-05 for Computers A, B, C, and D}
\format{A data frame with 1385 observations of 157 variables.}
\description{
  Example Wheel Running Data Wheel Running Data for
  2006-11-05 for Computers A, B, C, and D
}
\examples{
data(A061105)
str(A061105)
}
\keyword{datasets}

