\name{binning-package}
\alias{binning-package}
\alias{binning-package}
\title{Binning Wheel Running Data...}
\description{Binning Wheel Running Data}
\details{Functions to aggregate wheel running.

\tabular{ll}{
Package: \tab binning\cr
Type: \tab Package\cr
Version: \tab 0.04-3\cr
Date: \tab 2011-04-05\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
LazyData: \tab yes\cr
}}
\alias{binning-package}
\alias{binning}
\docType{package}
\author{Kevin Middleton}
\keyword{package}

