\docType{package}
\name{binning-package}
\alias{binning}
\alias{binning-package}
\title{Binning Wheel Running Data}
\description{
  Functions to aggregate wheel running.
}
\details{
  \tabular{ll}{ Package: \tab binning\cr Type: \tab
  Package\cr Version: \tab 0.06-1\cr Date: \tab
  2012-01-07\cr License: \tab GPL-2\cr LazyLoad: \tab
  yes\cr LazyData: \tab yes\cr }
}
\author{
  Kevin Middleton
}
\keyword{package}

