\name{binning-package}
\alias{binning-package}
\alias{binning}
\docType{package}

\title{Binning Wheel Running Data}

\description{Functions to aggregate wheel running.}

\details{
\tabular{ll}{
Package: \tab binning\cr
Type: \tab Package\cr
Version: \tab 0.04-1\cr
Date: \tab 2011-01-30\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
LazyData: \tab yes\cr
}
 An overview of how to use the package, including the most important functions 
}
\author{Kevin Middleton}

%\references{
% Literature or other references for background information 
%}

\keyword{ package }

%\seealso{
% Optional links to other man pages, e.g. 
% \code{\link[<pkg>:<pkg>-package]{<pkg>}} 
%}

%\examples{
%}
