\name{drop_dupes}
\alias{drop_dupes}
\title{Drop duplicated rows}
\usage{
  drop_dupes(dat)
}
\arguments{
  \item{dat}{data.frame from which to drop duplicated rows}
}
\value{
  data.frame without duplicated sensor recordings
}
\description{
  Drop duplicated rows
}
\details{
  Remove rows with duplicated TimeStamp
}
\author{
  Kevin Middleton
}

