\name{plot_running}
\alias{plot_running}
\title{Plot wheel running (new system)}
\usage{
  plot_running(data, save = FALSE)
}
\arguments{
  \item{data}{Data.frame returned from
  \code{\link{analyze_running}}.}

  \item{save}{Boolean (defulat = FALSE) should plots be
  saved?}
}
\description{
  Plot wheel running (new system)
}
\details{
  Create diagnostic plots for wheel running data,
}
\author{
  Kevin Middleton
}

