\name{read.dat}
\alias{read.dat}
\title{Read Wheel Running Data}
\usage{read.dat(file, tab.delim=FALSE)
}
\description{Read Wheel Running Data}
\details{Read either a comma-separated (.DAT) or tab-delimited (.txt) wheel running
data file.}
\value{A data.frame of wheel running data.}
\author{Kevin Middleton (kmm@csusb.edu)}
\keyword{data}
\arguments{\item{file}{either a comma-separated (.DAT) or tab-delimited (.txt) wheel
running data file}
\item{tab.delim}{Logical for whether the file is tab delimited or not. Default is FALSE,
because unmodified DAT files are comma-separated.}
}

