\name{stackdata}
\alias{stackdata}
\title{Stack Data Frames}
\usage{
  stackdata(...)
}
\arguments{
  \item{\dots}{data.frames to \code{rbind}}
}
\value{
  A data.frame with \dots stacked.
}
\description{
  Stack data frames on top of one another.
}
\details{
  Take an arbitrary number of aggregated data.frames in and
  return them as a single data.frame.
}
\author{
  Kevin Middleton (middletonk@missouri.edu)
}
\seealso{
  \code{\link{rbind}}
}
\keyword{data}

